+++ tests/files/listFilesWithNodeFs-flatignores/eslint.config.js
@@ -1,0 +1,3 @@
+module.exports = [
+  { ignores: ['ignored-dir/**', 'ignored-file.js'] },
+];